requirements.txt
altair==6.0.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
boto3==1.42.67
botocore==1.42.67
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fonttools==4.62.1
great_expectations==1.15.0
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.2.0
narwhals==2.18.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==26.0
pandas==2.2.2
parso==0.8.6
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==14.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
ruamel.yaml==0.19.1
s3transfer==0.16.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
wcwidth==0.6.0
